requirements.txt
requests
beautifulsoup4
nltk
gensim
cohere
gradio
numpy
matplotlib
pandas
scikit-learn
python-dotenv